+++ src/app/item-quiz/(components)/PlayingDisplay.tsx
@@ -2,17 +2,24 @@ import React from 'react'
 
 import { QuizItemButton } from '@/app/item-quiz/(components)/QuizItemButton'
 import { TOTAL_CHOICES } from '@/app/item-quiz/constants'
-import { QuizItem } from '@/app/item-quiz/types'
+import { LayoutPreference, QuizItem } from '@/app/item-quiz/types'
 import { getArrayOfLength } from '@/features/build/lib/getArrayOfLength'
+import { cn } from '@/lib/classnames'
 
 interface Props {
-  questionsForUI: QuizItem[]
   correctItemName: string
+  layoutPreference: LayoutPreference
+  questionsForUI: QuizItem[]
   onAnswerQuestion: (id: string) => void
 }
 
 export const PlayingDisplay = React.memo(
-  ({ correctItemName, questionsForUI, onAnswerQuestion }: Props) => (
+  ({
+    correctItemName,
+    layoutPreference,
+    questionsForUI,
+    onAnswerQuestion,
+  }: Props) => (
     <>
       <div
         id="item-name-container"
@@ -23,15 +30,39 @@ export const PlayingDisplay = React.memo(
       </div>
 
       {/** Mobile Grid */}
-      <div className="grid grid-cols-2 gap-4 sm:hidden">
+      <div
+        className={cn(
+          'grid grid-cols-2 gap-4',
+          layoutPreference === 'desktop' && 'sm:hidden',
+        )}
+      >
         {getArrayOfLength(TOTAL_CHOICES).map((_, index) => {
+          let cssOrder: string | undefined = undefined
+          if (layoutPreference === 'mobile') {
+            switch (index) {
+              case 0:
+                cssOrder = 'order-1'
+                break
+              case 1:
+                cssOrder = 'order-3'
+                break
+              case 2:
+                cssOrder = 'order-4'
+                break
+              case 3:
+                cssOrder = 'order-2'
+                break
+            }
+          }
+
           return (
-            <QuizItemButton
-              key={questionsForUI[index].id}
-              item={questionsForUI[index]}
-              itemIndex={index}
-              onClick={() => onAnswerQuestion(questionsForUI[index].id)}
-            />
+            <div key={questionsForUI[index].id} className={cn(cssOrder)}>
+              <QuizItemButton
+                item={questionsForUI[index]}
+                itemIndex={index}
+                onClick={() => onAnswerQuestion(questionsForUI[index].id)}
+              />
+            </div>
           )
         })}
       </div>
@@ -39,7 +70,10 @@ export const PlayingDisplay = React.memo(
       {/** Desktop grid */}
       <div
         id="quiz-choice-grid"
-        className="hidden sm:grid sm:grid-cols-3 sm:gap-4"
+        className={cn(
+          'hidden sm:grid sm:grid-cols-3 sm:gap-4',
+          layoutPreference === 'mobile' && 'sm:hidden',
+        )}
       >
         {/** Left arrow or 3 Key */}
         <div className="col-span-1 flex w-full items-center justify-center">